Bổ sung những hàm kiểm tra các ID có tồn tại trong table không.

+++ trunk/RealEstateBusinessLogicObject/AddressBLO.cs
@@ -32,9 +32,30 @@ namespace RealEstateBusinessLogicObject
         /// <returns>ID of row just insert</returns>
         public override int Insert(RealEstateDataContext.ADDRESS entity)
         {
-            entity.ID = _db.CreateID();
-            _db.Insert(entity);
-            return entity.ID;
+            if ((new RealEstateDataAccessObject.NationDAO()).ValidationID(entity.NationID))
+            {
+                if (new RealEstateDataAccessObject.CityDAO().ValidationID(entity.CityID))
+                {
+                    if (new RealEstateDataAccessObject.DistrictDAO().ValidationID((int)entity.DistrictID) || entity.DistrictID == null)
+                    {
+                        if (new RealEstateDataAccessObject.WardDAO().ValidationID((int)entity.WardID) || entity.WardID == null)
+                        {
+                            if (new RealEstateDataAccessObject.StreetDAO().ValidationID((int)entity.StreetID) || entity.StreetID == null)
+                            {
+                                entity.ID = _db.CreateID();
+                                _db.Insert(entity);
+                                return entity.ID;
+                            }
+                            else throw new RealEstateDataContext.Utility.StreetID();
+                        }
+                        else throw new RealEstateDataContext.Utility.WardID();
+                    }
+                    else throw new RealEstateDataContext.Utility.DistrictID();
+                }
+                else throw new RealEstateDataContext.Utility.CityID();
+            }
+            else throw new RealEstateDataContext.Utility.NationID();
+            
         }
 
         /// <summary>
@@ -47,19 +68,39 @@ namespace RealEstateBusinessLogicObject
         /// <param name="streetID">ID of street</param>
         /// <param name="detail">Detail Address</param>
         /// <returns>ID of row just insert</returns>
-        public int Insert(int nationID, int cityID, int districtID, int wardID, int streetID, string detail)
+        public int Insert(int nationID, int cityID, int? districtID, int? wardID, int? streetID, string detail)
         {
-            RealEstateDataContext.ADDRESS entity = new RealEstateDataContext.ADDRESS();
-            entity.ID = _db.CreateID();
-            entity.NationID = nationID;
-            entity.CityID = cityID;
-            entity.DistrictID = districtID;
-            entity.WardID = wardID;
-            entity.StreetID = streetID;
-            entity.Detail = detail;
+            if ((new RealEstateDataAccessObject.NationDAO()).ValidationID(nationID))
+            {
+                if (new RealEstateDataAccessObject.CityDAO().ValidationID(cityID))
+                {
+                    if (new RealEstateDataAccessObject.DistrictDAO().ValidationID((int)districtID) || districtID == null)
+                    {
+                        if (new RealEstateDataAccessObject.WardDAO().ValidationID((int)wardID) || wardID == null)
+                        {
+                            if (new RealEstateDataAccessObject.StreetDAO().ValidationID((int)streetID) || streetID == null)
+                            {
+                                RealEstateDataContext.ADDRESS entity = new RealEstateDataContext.ADDRESS();
+                                entity.ID = _db.CreateID();
+                                entity.NationID = nationID;
+                                entity.CityID = cityID;
+                                entity.DistrictID = districtID;
+                                entity.WardID = wardID;
+                                entity.StreetID = streetID;
+                                entity.Detail = detail;
 
-            _db.Insert(entity);
-            return entity.ID;
+                                _db.Insert(entity);
+                                return entity.ID;
+                            }
+                            else throw new RealEstateDataContext.Utility.StreetID();
+                        }
+                        else throw new RealEstateDataContext.Utility.WardID();
+                    }
+                    else throw new RealEstateDataContext.Utility.DistrictID();
+                }
+                else throw new RealEstateDataContext.Utility.CityID();
+            }
+            else throw new RealEstateDataContext.Utility.NationID();         
         }
 
         /// <summary>
@@ -69,8 +110,32 @@ namespace RealEstateBusinessLogicObject
         /// <returns>ID of row just update</returns>
         public override int Update(RealEstateDataContext.ADDRESS entity)
         {
-            _db.Update(entity);
-            return entity.ID;
+            if (ValidationID(entity.ID))
+            {
+                if (new RealEstateDataAccessObject.NationDAO().ValidationID(entity.NationID))
+                {
+                    if (new RealEstateDataAccessObject.CityDAO().ValidationID(entity.CityID))
+                    {
+                        if (new RealEstateDataAccessObject.DistrictDAO().ValidationID((int)entity.DistrictID) || entity.DistrictID == null)
+                        {
+                            if (new RealEstateDataAccessObject.WardDAO().ValidationID((int)entity.WardID) || entity.WardID == null)
+                            {
+                                if (new RealEstateDataAccessObject.StreetDAO().ValidationID((int)entity.StreetID) || entity.StreetID == null)
+                                {
+                                    _db.Update(entity);
+                                    return entity.ID;
+                                }
+                                else throw new RealEstateDataContext.Utility.StreetID();
+                            }
+                            else throw new RealEstateDataContext.Utility.WardID();
+                        }
+                        else throw new RealEstateDataContext.Utility.DistrictID();
+                    }
+                    else throw new RealEstateDataContext.Utility.CityID();
+                }
+                else throw new RealEstateDataContext.Utility.NationID();
+            }
+            else throw new RealEstateDataContext.Utility.AddressID();
         }
 
         /// <summary>
@@ -84,19 +149,43 @@ namespace RealEstateBusinessLogicObject
         /// <param name="streetID">ID of street</param>
         /// <param name="detail">Detail Address</param>
         /// <returns>ID of row just update</returns>
-        public int Update(int id, int nationID, int cityID, int districtID, int wardID, int streetID, string detail)
+        public int Update(int id, int nationID, int cityID, int? districtID, int? wardID, int? streetID, string detail)
         {
-            RealEstateDataContext.ADDRESS entity = new RealEstateDataContext.ADDRESS();
-            entity.ID = id;
-            entity.NationID = nationID;
-            entity.CityID = cityID;
-            entity.DistrictID = districtID;
-            entity.WardID = wardID;
-            entity.StreetID = streetID;
-            entity.Detail = detail;
+            if (ValidationID(id))
+            {
+                if (new RealEstateDataAccessObject.NationDAO().ValidationID(nationID))
+                {
+                    if (new RealEstateDataAccessObject.CityDAO().ValidationID(cityID))
+                    {
+                    if (new RealEstateDataAccessObject.DistrictDAO().ValidationID((int)districtID) || districtID == null)
+                    {
+                        if (new RealEstateDataAccessObject.WardDAO().ValidationID((int)wardID) || wardID == null)
+                        {
+                            if (new RealEstateDataAccessObject.StreetDAO().ValidationID((int)streetID) || streetID == null)
+                            {
+                                RealEstateDataContext.ADDRESS entity = new RealEstateDataContext.ADDRESS();
+                                entity.ID = id;
+                                entity.NationID = nationID;
+                                entity.CityID = cityID;
+                                entity.DistrictID = districtID;
+                                entity.WardID = wardID;
+                                entity.StreetID = streetID;
+                                entity.Detail = detail;
 
-            _db.Update(entity);
-            return entity.ID;
+                                _db.Update(entity);
+                                return entity.ID;
+                            }
+                            else throw new RealEstateDataContext.Utility.StreetID();
+                        }
+                        else throw new RealEstateDataContext.Utility.WardID();
+                    }
+                    else throw new RealEstateDataContext.Utility.DistrictID();
+                    }
+                    else throw new RealEstateDataContext.Utility.CityID();
+                }
+                else throw new RealEstateDataContext.Utility.NationID();
+            }
+            else throw new RealEstateDataContext.Utility.AddressID();
         }
 
 
@@ -104,24 +193,23 @@ namespace RealEstateBusinessLogicObject
         /// Delete a row from ADDRESS table
         /// </summary>
         /// <param name="ID">ID of row want to delete</param>
+        /// <exception cref="AddressID"></exception>
         public override void Delete(int ID)
         {
-            _db.Delete(ID);
+            if (ValidationID(ID))
+            {
+                _db.Delete(ID);
+            }
+            else throw new RealEstateDataContext.Utility.AddressID();
         }
 
-        /// <summary>
-        /// Get a row in ADDRESS table
-        /// </summary>
-        /// <param name="ID">ID of row</param>
-        /// <returns>Entity</returns>
         public override RealEstateDataContext.ADDRESS GetARecord(int ID)
         {
-            return _db.GetARecord(ID);
-        }
-
-        public RealEstateDataContext.NATION GetNation(int ID)
-        {
-            return _db.GetARecord(ID).NATION;
+            if (ValidationID(ID))
+            {
+                return _db.GetARecord(ID);
+            }
+            else throw new RealEstateDataContext.Utility.AddressID();
         }
     }
 }
